Setting up the framework for the vision system being able to visually detect other robots.
We are currently missing the Mobilenet model for actually detecting robots. In addition to this, robotWorld is completely useless right now.
It is basically does nothing except exist and provide a stub for visionComponent. A new message in clad has been created, RobotObservedRobot but this is never sent anywhere yet.
A new VisionMode has also been made, but it is never activated anywhere.

That should be all the changes I made, but I probably forgot some. Anyway, more work will be done once the Mobilenet model has been created.